Add list_feature_groups API (#647)
